3d mesh volume & surface area calculation; implemented -i flag

+++ src/headers/stl.h
@@ -1,6 +1,6 @@
 #ifndef STL_H
 #define STL_H
 
-void getSTLParams(char* fname, char type, int params[4]);
+void getSTLParams(char* fname, char type, float params[4]);
 
 #endif
